Refactor fast encoding check to automatically select the fastest available implementation
