+++ benchmark/string_coderange_scan.rb
@@ -1,4 +1,4 @@
-str = 'abc€›ﬁ!‰,' * 1_000_000
+str = 'abc€›ﬁ!‰,' * 1_000
 1_000.times do
   str.force_encoding(Encoding::UTF_8) # clear coderange
   str.valid_encoding?
